Factor out common object value transpiration logic
